+++ R/build.r
@@ -42,7 +42,11 @@ build <- function(pkg = ".", path = NULL, binary = FALSE, vignettes = TRUE,
       paste0(args, collapse = " "))
     ext <- if (.Platform$OS.type == "windows") "zip" else "tgz"
   } else {
-    args <- c(args, "--no-manual", "--no-resave-data")
+    args <- c(args, "--no-resave-data")
+
+    if (!("--manual" %in% args)) {
+      args <- c(args, "--no-manual")
+    }
 
     if (!vignettes) {
       args <- c(args, "--no-build-vignettes")