+++ NEWS.md
@@ -3,6 +3,12 @@
 # devtools 1.6.1
 
 * Don't set non-portable compiler flags on Solaris.
+## The release process
+
+* `release()` gains two new arguments `args` and `build_args` that are passed
+  to `check()` and `build()` (@krlmlr, #603).
+
+## Bug fixes
 
 * The file `template.Rproj` is now correctly installed and the function
   `use_rstudio` works as it should. (#595, @hmalmedal)